Removed nullability where not appropriate
The Tavis API makes somewhat fuzzy use of the term "parameter", and in some cases a parameter value is allowed to be null, but in cases where it reports the 'parameters' that are present in a particular URL as interpretted through a particular template, it will never report null values. (It provides a dictionary containing only those values that were present, so there's no occasion for this to have null values.)

This was not reflected in the API - it used IDictionary<string, object?> even in cases where the value would not be null. This change corrects that.